Update dependencies for Python3.10 as installed on ubuntu 22.04
The bluez version 0.23 is incombatible to setuptools >=58.0.0.
Ubuntu 22.04 ships 59.6.0 which leads to pybluez not building
anymore.

Furthermore mkdirs is no longer available on modern python3 including
the latest python 3.10.5. We therefore replace it with makedirs.

As a last step we need pyyaml in our requirements as it is not installed
by default into the python distributions.

We document the steps to create a virtual environment in the README for
the next person to find themself with a modern python (¬_¬)

Signed-off-by: Mimoja <git@mimoja.de>