build(deps): pin first pybluez version that is compatible with modern SetupTools
PyBluez is already aware of the issue that SetupTools no longer
supports "use_2to3".
https://github.com/pybluez/pybluez/issues/413
And fixed already in September 2021 in:
https://github.com/pybluez/pybluez/commit/37d78880179b2a83e7052e0c2b9393499dd3b857
Sadly there is currently no new pypi release since 2019 so there is no
tagged release for us to upgrade to.
https://github.com/pybluez/pybluez/issues/416
This is due to the lack of a pybluez maintainer
https://github.com/pybluez/pybluez/issues/411

To solve this we are pinning the commit that removed the use_2to3
in the requirements.txt allowing us to build with moden python distributions.

Testing done: Printed one Label with text and spacing

We are also only pulling in pybluez on linux as it wont work on windows and macos.
Fixes: 20f11b082eeafc6367c0043c6c5dac8c27bc2a4b Transition to qt6 (#29)

Signed-off-by: Mimoja <git@mimoja.de>